Extract pointer/keyboard event handlers into eventHandlers.ts
Move the 12 mouse/pointer handlers (doMouseDown/Up/Move, doClick,
doRightClick, doMiddleClick, doDoubleClick, doHover, doWheel, and the
three measure-mode variants) plus handleKeyDown/handleKeyUp out of
MGWebGL. They were already written in collaborator style — registered
as `self.doMouseDown(evt, self)` on the canvas — so they become free
functions taking the live instance as `self: MGWebGL`. All interaction
state and the helper methods they call (getAtomFomMouseXY, updateLabels,
getThreeWayMatrixAndViewPort, drawScene, …) stay on the instance,
reached through `self`; the class methods become thin delegators.

Typing `self` as MGWebGL (the originals took an untyped `self`) surfaced
a few latent mismatches against mistyped .d.ts entries
(keyboardAccelerators/keysDown) — handled with narrow, behaviour-
preserving casts documented in place. getOffsetRect is now exported and
hoverDebounceTimeout un-privated so the collaborator can reach them.

This is the seam a future gesture-recognizer library would sit above:
it would consume normalized input and drive these as the action verbs.

Verified: project typechecks clean; an independent line-by-line diff of
all 14 handlers against HEAD confirms behaviourally identical output.
